Update templates to not use aliases

+++ templates/math_common.cpp
@@ -5,9 +5,9 @@ constexpr int gcd(int a, int b) {
 }
 constexpr int lcm(int a, int b) { return (a * b) / gcd(a, b); }
 
-constexpr ll binpow(ll a, ll b, ll m) {
+constexpr long long binpow(long long a, long long b, long long m) {
   a %= m;
-  ll res = 1;
+  long long res = 1;
   while (b > 0) {
     if (b & 1)
       res = res * a % m;
