+++ templates/mod_int.cpp
@@ -1,13 +1,13 @@
 struct mint {
-  ll v{};
+  long long v{};
   mint() = default;
-  mint(ll vx) : v((vx % M + M) % M) {}
+  mint(long long vx) : v((vx % M + M) % M) {}
   friend mint operator+(const mint &a, const mint &b) { return {(a.v + b.v) % M}; }
-  friend mint operator+(ll a, const mint &b) { return {(a + b.v) % M}; }
-  friend mint operator+(const mint &a, ll b) { return {(a.v + b) % M}; }
+  friend mint operator+(long long a, const mint &b) { return {(a + b.v) % M}; }
+  friend mint operator+(const mint &a, long long b) { return {(a.v + b) % M}; }
   friend mint operator*(const mint &a, const mint &b) { return {(a.v * b.v) % M}; }
-  friend mint operator*(ll a, const mint &b) { return {(a * b.v) % M}; }
-  friend mint operator*(const mint &a, ll b) { return {(a.v * b) % M}; }
+  friend mint operator*(long long a, const mint &b) { return {(a * b.v) % M}; }
+  friend mint operator*(const mint &a, long long b) { return {(a.v * b) % M}; }
   friend bool operator==(const mint &a, const mint &b) { return a.v != b.v; }
   friend bool operator<(const mint &a, const mint &b) { return a.v < b.v; }
   friend bool operator>(const mint &a, const mint &b) { return a.v > b.v; }
